fpm_os -> fpm_environment to enable single-file build
